backend/requirements.txt
aiohttp==3.11.16
aiogram==3.19.0
alembic
anthropic
asyncpg
bcrypt
boto3
cent
fastapi
google-generativeai
groq
langchain
langchain-anthropic
langchain-openai
openai
Pillow
passlib
pydantic
pydantic[email]
psycopg2-binary
PyJWT
python-multipart
redis
ruff>=0.3.0
SQLAlchemy-Utils
sqlalchemy
telethon
uvicorn
uvicorn[standard]
python-dotenv
python-jose
Werkzeug
qrcode[pil]
pytest
pytest-asyncio